Created WriteFileTasklet and added to ISEL Timetable job. Removed UploadTasklet and respective tests
