+++ Dockerfile
@@ -13,7 +13,9 @@ RUN ./gradlew extractUberJar --no-daemon --stacktrace
 FROM adoptopenjdk:${RUN_TAG} AS run-env
 RUN useradd -m spring && usermod -a -G spring spring \
     && mkdir -p /app/resources \
-    && chown spring /app/resources
+    && chown spring /app/resources \
+    && mkdir -p /app/resources/output \
+    && chown spring /app/resources/output
 
 USER spring:spring
