+++ src/main/resources/application.properties
@@ -17,6 +17,7 @@ ion.core-token=${ION_CORE_TOKEN}
 ion.core-request-timeout-seconds=5
 ion.resources-folder=/app/resources
 ion.core-retries=3
+ion.local-file-output-folder=/app/resources/output
 
 # spring mail
 spring.mail.host=smtp.gmail.com
