+++ src/test/resources/application.properties
@@ -24,7 +24,5 @@ email.sender=alert-mailbox@domain.com
 ion.core-base-url=test
 ion.core-token=test
 ion.core-request-timeout-seconds=1
-
-
-
 ion.resources-folder=src/test/resources/
+ion.local-file-output-folder=src/test/resources/output/
